Remove generic ParseNumber since we only ever parse uint (#3)
I build XAPerfTestRunner with NativeAOT and I found out that ParseNumber<T> did not work.
It seems like at some point there was a need for a generic solution, but it doesn't seem to
be needed anymore. I suggest simplifying the parsing and also making the -r argument
codepath compatible with NativeAOT.